requirements.txt
# FastAPI Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-multipart==0.0.6

# Database
sqlalchemy==2.0.36
pyodbc==5.3.0
python-dotenv==1.0.0

# Data Processing
pandas==2.2.3
numpy==2.1.2

# File Handling
aiofiles==23.2.0

# Validation
pydantic==2.8.2

# Templates
jinja2==3.1.2

# HTTP Client
requests==2.31.0